start adicionado

+++ src/controllers/task.controller.js
@@ -1,5 +1,9 @@
 const TaskModel = require("./../models/task.model");
-
+const {
+    notFoundError,
+    ObjectIdCastIdError,
+} = require("./../errors/mongodb.erros");
+const mongoose = require("mongoose");
 class TaskController {
     constructor(req, res) {
         this.req = req;
@@ -23,17 +27,22 @@ class TaskController {
 
             this.res.status(201).send(newTask);
         } catch (err) {
-            this.res.status(500).send(error.message);
+            this.res.status(500).send(err.message);
         }
     }
 
     async getById() {
         try {
             const taskId = this.req.params.id;
+
+            if (!mongoose.Types.ObjectId.isValid(taskId)) {
+                return ObjectIdCastIdError(this.res);
+            }
+
             const tasks = await TaskModel.findById(taskId);
 
             if (!tasks) {
-                return res.status(404).send("Essa tarefa não foi encontrada");
+                return notFoundError(this.res);
             }
 
             this.res.status(200).send(tasks);
@@ -47,7 +56,14 @@ class TaskController {
             const taskId = this.req.params.id;
             const taskData = this.req.body;
 
+            if (!mongoose.Types.ObjectId.isValid(taskId)) {
+                return ObjectIdCastIdError(this.res);
+            }
             const taskToUpdate = await TaskModel.findById(taskId);
+
+            if (!taskToUpdate) {
+                return notFoundError(this.res);
+            }
 
             const allowedUpdates = ["isCompleted"];
 
@@ -72,10 +88,14 @@ class TaskController {
         try {
             const taskId = this.req.params.id;
 
+            if (!mongoose.Types.ObjectId.isValid(taskId)) {
+                return ObjectIdCastIdError(this.res);
+            }
+
             const taskToDeleted = await TaskModel.findById(taskId);
 
             if (!taskToDeleted) {
-                return this.res.status(404).send("Tarefa não encontrada");
+                return notFoundError(this.res);
             }
 
             const deletedTask = await TaskModel.findByIdAndDelete(taskId);
